updateSettings: migrate to /authManagement, add req and res types, no-verify due to logic in password checks
